v68k:  Cure undefined behavior in microcode_DIVS()
We were shifting data (remainder's sign bit) into the result's sign bit,
which is undefined if the incoming bit is non-zero.  Make it an unsigned
operation instead, whose result has no sign bit, thus solving the issue.

+++ 68k/libs/v68k/v68k/microcode.cc
@@ -1079,7 +1079,8 @@ namespace v68k
 			//const int16_t remainder = dividend % divisor;
 			const int16_t remainder = dividend - quotient * divisor;
 			
-			pb.result = remainder << 16 | uint16_t( quotient );
+			pb.result = (uint16_t) remainder << 16
+			          | (uint16_t) quotient;
 		}
 		
 		return Ok;
